+++ code/renderercommon/tr_image_png.c
@@ -532,7 +532,7 @@ static uint32_t DecompressIDATs(struct BufferedFile *BF, uint8_t **Buffer)
 
 	if(!(BF && Buffer))
 	{
-		return(-1);
+		return((uint32_t)-1);
 	}
 
 	/*
@@ -553,7 +553,7 @@ static uint32_t DecompressIDATs(struct BufferedFile *BF, uint8_t **Buffer)
 
 	if(!FindChunk(BF, PNG_ChunkType_IDAT))
 	{
-		return(-1);
+		return((uint32_t)-1);
 	}
 
 	/*
@@ -576,7 +576,7 @@ static uint32_t DecompressIDATs(struct BufferedFile *BF, uint8_t **Buffer)
 
 			BufferedFileRewind(BF, BytesToRewind);
 
-			return(-1);
+			return((uint32_t)-1);
 		}
 
 		/*
@@ -613,7 +613,7 @@ static uint32_t DecompressIDATs(struct BufferedFile *BF, uint8_t **Buffer)
 			{
 				BufferedFileRewind(BF, BytesToRewind);
 
-				return(-1);
+				return((uint32_t)-1);
 			}
 
 			BytesToRewind += Length + PNG_ChunkCRC_Size;
@@ -626,7 +626,7 @@ static uint32_t DecompressIDATs(struct BufferedFile *BF, uint8_t **Buffer)
 	CompressedData = ri.Malloc(CompressedDataLength);
 	if(!CompressedData)
 	{
-		return(-1);
+		return((uint32_t)-1);
 	}
 
 	CompressedDataPtr = CompressedData;
@@ -646,7 +646,7 @@ static uint32_t DecompressIDATs(struct BufferedFile *BF, uint8_t **Buffer)
 		{
 			ri.Free(CompressedData); 
 
-			return(-1);
+			return((uint32_t)-1);
 		}
 
 		/*
@@ -680,14 +680,14 @@ static uint32_t DecompressIDATs(struct BufferedFile *BF, uint8_t **Buffer)
 			{
 				ri.Free(CompressedData); 
 
-				return(-1);
+				return((uint32_t)-1);
 			}
 
 			if(!BufferedFileSkip(BF, PNG_ChunkCRC_Size))
 			{
 				ri.Free(CompressedData); 
 
-				return(-1);
+				return((uint32_t)-1);
 			}
 
 			memcpy(CompressedDataPtr, OrigCompressedData, Length);
@@ -718,7 +718,7 @@ static uint32_t DecompressIDATs(struct BufferedFile *BF, uint8_t **Buffer)
 	{
 		ri.Free(CompressedData);
 
-		return(-1);
+		return((uint32_t)-1);
 	}
 
 	/*
@@ -730,7 +730,7 @@ static uint32_t DecompressIDATs(struct BufferedFile *BF, uint8_t **Buffer)
 	{
 		ri.Free(CompressedData);
 
-		return(-1);
+		return((uint32_t)-1);
 	}
 
 	/*
@@ -761,7 +761,7 @@ static uint32_t DecompressIDATs(struct BufferedFile *BF, uint8_t **Buffer)
 	{
 		ri.Free(DecompressedData);
 
-		return(-1);
+		return((uint32_t)-1);
 	}
 
 	/*
@@ -2364,7 +2364,7 @@ void R_LoadPNG(const char *name, byte **pic, int *width, int *height)
 	 *  Rewind to the start of the file.
 	 */
 
-	if(!BufferedFileRewind(ThePNG, -1))
+	if(!BufferedFileRewind(ThePNG, (unsigned)-1))
 	{
 		CloseBufferedFile(ThePNG);
